+++ divine-modules/divine-warehouse/src/main/java/com/divine/warehouse/domain/entity/ShipmentOrder.java
@@ -33,7 +33,7 @@ public class ShipmentOrder extends BaseOrder {
     /**
      * 入库类型
      */
-    private Long optType;
+    private Integer optType;
 
     /**
      * 第三方订单号
